Fixes #1391 - Provide a simplistic simulation of mesos for scale testing
Create a new sub module to separate simulation code from production code.
Also use a guice provider to instantiate the driver so that
we could substitute the provider to instatiate a simulated driver.
There were some global variable references that had to be removed for this to work.
Fixes #1391 - Provide a simplistic simulation of mesos for scale testing

Create a new sub module to separate simulation code from production code.
Also use a guice provider to instantiate the driver so that
we could substitute the provider to instatiate a simulated driver.
There were some global variable references that had to be removed for this to work.